refactoring of game logic
